# feat(server): faster folder responses