--- a/server/src/interfaces/asset.interface.ts
+++ b/server/src/interfaces/asset.interface.ts
@@ -149,6 +149,7 @@ export const IAssetRepository = 'IAssetRepository';
 export interface IAssetRepository {
   getAssetsByOriginalPath(userId: string, partialPath: string): Promise<AssetEntity[]>;
   getUniqueOriginalPaths(userId: string): Promise<string[]>;
+  removeEmptyFolders(): Promise<void>;
   create(asset: AssetCreate): Promise<AssetEntity>;
   getByIds(
     ids: string[],