--- a/server/src/services/job.service.ts
+++ b/server/src/services/job.service.ts
@@ -212,6 +212,7 @@ export class JobService {
       { name: JobName.USER_SYNC_USAGE },
       { name: JobName.QUEUE_FACIAL_RECOGNITION, data: { force: false, nightly: true } },
       { name: JobName.CLEAN_OLD_SESSION_TOKENS },
+      { name: JobName.CLEAN_FOLDER_TABLE },
     ]);
   }