--- a/web/src/lib/components/shared-components/tree/tree-items.svelte
+++ b/web/src/lib/components/shared-components/tree/tree-items.svelte
@@ -1,6 +1,6 @@
 <script lang="ts">
   import Tree from '$lib/components/shared-components/tree/tree.svelte';
-  import { normalizeTreePath, type RecursiveObject } from '$lib/utils/tree-utils';
+  import { joinPaths, type RecursiveObject } from '$lib/utils/tree-utils';
 
   export let items: RecursiveObject;
   export let parent = '';
@@ -12,12 +12,14 @@
 
 <ul class="list-none ml-2">
   {#each Object.entries(items) as [path, tree]}
-    {@const value = normalizeTreePath(`${parent}/${path}`)}
-    {@const key = value + getColor(value)}
-    {#key key}
-      <li>
-        <Tree {parent} value={path} {tree} {icons} {active} {getLink} {getColor} />
-      </li>
-    {/key}
+    {#if path !== '\0'}
+      {@const value = joinPaths(parent, path)}
+      {@const key = value + getColor(value)}
+      {#key key}
+        <li>
+          <Tree {parent} value={path} {tree} {icons} {active} {getLink} {getColor} />
+        </li>
+      {/key}
+    {/if}
   {/each}
 </ul>