… server/src/services/job.service.spec.ts
… server/test/repositories/asset.repository.mock.ts
… web/src/lib/stores/folders.store.ts
… web/src/lib/utils/tree-utils.ts
… web/src/routes/(user)/folders/[[photos=photos]]/[[assetId=id]]/+page.svelte
… web/src/routes/(user)/tags/[[photos=photos]]/[[assetId=id]]/+page.svelte