@alextran1502 2024-09-01
If you want to update the folders instantly, do we have a job to do that?