@mertalev 2024-09-01
It doesn't handle updates that change `originalPath` right now (so mainly storage template migration). That will just be modifying the update query and won't be tied to a particular job.