@jrasm91 2024-09-01
This seems to add a lot of edge case handling in the frontend. Can we simplify it by handling more of them server-side?