@mertalev 2024-09-01
We can remove `normalizeTreePath` since the server response will never have trailing slashes. They won't happen in code either as long as `joinPaths` is used instead of basic string concatenation.