@jrasm91 2024-09-01
Do you mind moving the vacuum logic into a separate pull request?